pyproject.toml
[tool.poetry]
name = "botjam mod"
version = "0.1.0"
description = ""
authors = ["timelessnesses <mooping3roblox@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = {git = "https://github.com/Rapptz/discord.py", extras = ["speed"]}
aiohttp = {extras = ["speedups"], version = "^3.8.1"}
uvloop = {version = "^0.16.0", platform = "linux"}
aiofiles = "^0.8.0"
jishaku = "^2.4.0"
psutil = "^5.9.0"
python-dotenv = "^0.20.0"
python-dateutil = "^2.8.2"
parsedatetime = "^2.6"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}
ez-aiohttp = "^0.1.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
